Separate error types into their own file
