+++ zsh-module/src/types/mod.rs
@@ -1,2 +1,3 @@
 pub mod cstring;
+pub mod error;
 pub mod files;
